refactor: move folders - create public folder for all exported components - refactor Configure folder to have content folder - move ErrorTextBox to higher level components folder
