+++ src/components/Configure/content/README.md
@@ -1,0 +1,3 @@
+### content folder
+This folder contains views and components related to the right content panel.This folder
+separates components from the nav sidebar on the left (ObjectManagementNav).
