+++ src/components/Configure/content/fields/ReadFields.tsx
@@ -1,0 +1,13 @@
+import { RequiredFieldMappings } from './FieldMappings';
+import { OptionalFields } from './OptionalFields';
+import { RequiredFields } from './RequiredFields';
+
+export function ReadFields() {
+  return (
+    <>
+      <RequiredFields />
+      <RequiredFieldMappings />
+      <OptionalFields />
+    </>
+  );
+}
