--- src/components/index.ts
@@ -1,3 +1,0 @@
-export * from './AmpersandProvider';
-export * from './Configure';
-export * from './Connect/ConnectProvider';
