+++ src/public/index.ts
@@ -1,0 +1,7 @@
+/**
+ * Contains all the features to be exported out of the library
+ */
+export * from '../components/AmpersandProvider';
+export * from '../components/Configure';
+export * from '../components/Connect/ConnectProvider';
+export * from '../hooks/useIsIntegrationInstalled';
